+++ lib/axn/extras.rb
@@ -1,3 +1,4 @@
 # frozen_string_literal: true
 
 require "axn/extras/strategies/client"
+require "axn/extras/strategies/vernier"